+++ library/src/plugins/official/backend/watchers/removeFragments.ts
@@ -3,7 +3,6 @@
 // Description: Remember, SSE is just a regular SSE request but with the ability to send 0-inf messages to the client.
 
 import {
-  DefaultFragmentsSettleDurationMs,
   DefaultFragmentsUseViewTransitions,
   EventTypes,
 } from '../../../../engine/consts'
@@ -14,7 +13,7 @@ import {
   docWithViewTransitionAPI,
   supportsViewTransitions,
 } from '../../../../utils/view-transtions'
-import { SWAPPING_CLASS, datastarSSEEventWatcher } from '../shared'
+import { datastarSSEEventWatcher } from '../shared'
 
 export const RemoveFragments: WatcherPlugin = {
   type: PluginType.Watcher,
@@ -24,8 +23,6 @@ export const RemoveFragments: WatcherPlugin = {
       EventTypes.RemoveFragments,
       ({
         selector,
-        settleDuration:
-          settleDurationRaw = `${DefaultFragmentsSettleDurationMs}`,
         useViewTransition:
           useViewTransitionRaw = `${DefaultFragmentsUseViewTransitions}`,
       }) => {
@@ -33,20 +30,13 @@ export const RemoveFragments: WatcherPlugin = {
           throw initErr('NoSelectorProvided', ctx)
         }
 
-        const settleDuration = Number.parseInt(settleDurationRaw)
         const useViewTransition = isBoolString(useViewTransitionRaw)
         const removeTargets = document.querySelectorAll(selector)
 
         const applyToTargets = () => {
           for (const target of removeTargets) {
-            target.classList.add(SWAPPING_CLASS)
+            target.remove()
           }
-
-          setTimeout(() => {
-            for (const target of removeTargets) {
-              target.remove()
-            }
-          }, settleDuration)
         }
 
         if (useViewTransition && supportsViewTransitions) {