+++ library/src/plugins/official/dom/attributes/on.ts
@@ -11,7 +11,7 @@ import {
   Requirement,
 } from '../../../../engine/types'
 import { tagHas, tagToMs } from '../../../../utils/tags'
-import { camel, modifyCasing } from '../../../../utils/text'
+import { camel, kebab, modifyCasing } from '../../../../utils/text'
 import { debounce, delay, throttle } from '../../../../utils/timing'
 import { supportsViewTransitions } from '../../../../utils/view-transtions'
 
@@ -108,7 +108,9 @@ export const On: AttributePlugin = {
       rafId = requestAnimationFrame(raf)
 
       return () => {
-        if (rafId) cancelAnimationFrame(rafId)
+        if (rafId) {
+          cancelAnimationFrame(rafId)
+        }
       }
     }
 
@@ -150,7 +152,10 @@ export const On: AttributePlugin = {
       callback = targetOutsideCallback
     }
 
-    const eventName = modifyCasing(key, mods)
+    // Default to kebab-case and allow modifying
+    let eventName = kebab(key)
+    eventName = modifyCasing(eventName, mods)
+
     target.addEventListener(eventName, callback, evtListOpts)
     return () => {
       target.removeEventListener(eventName, callback)