+++ library/src/vendored/fetch-event-source.ts
@@ -259,7 +259,6 @@ export interface FetchEventSourceInit extends RequestInit {
 }
 
 export function fetchEventSource(
-  ctx: RuntimeContext,
   input: RequestInfo,
   {
     signal: inputSignal,
@@ -360,14 +359,14 @@ export function fetchEventSource(
             retryInterval *= retryScaler // exponential backoff
             retryInterval = Math.min(retryInterval, retryMaxWaitMs)
             retries++
-            if (retries >= retryMaxCount) {
+            if (retries > retryMaxCount) {
               // we should not retry anymore:
               dispose()
-              // Max retries hit, check your server or network connection
-              reject(runtimeErr('SseMaxRetries', ctx, { retryMaxCount }))
+              // Max retries reached, check your server or network connection
+              reject('Max retries reached.')
             } else {
               console.error(
-                `Datastar failed to reach ${rest.method}: ${input.toString()} retry in ${interval}ms`,
+                `Datastar failed to reach ${input.toString()} retrying in ${interval}ms.`,
               )
             }
           } catch (innerErr) {