+++ sdk/clojure/src/dev/examples/data_dsl.clj
@@ -3,12 +3,14 @@
     [starfederation.datastar.clojure.consts :as consts]
     [starfederation.datastar.clojure.api :as d*]))
 
+;; Examples of how one might want to build a higher level api
+;; on top of the SDK
+
 (def example
   {:event ::merge-fragments
    :fragments ["<div>hello</div>"]
    d*/selector "foo"
    d*/merge-mode d*/mm-append
-   d*/settle-duration 500
    d*/use-view-transition true})
 
 
@@ -28,7 +30,6 @@
 (sse-event example)
 ; ["selector foo"
 ;  "mergeMode append"
-;  "settleDuration 500"
 ;  "useViewTransition true"
 ;  "fragments <div>hello</div>"]
 
@@ -57,7 +58,6 @@
 ; retry: 1000\n
 ; data: selector foo\n
 ; data: mergeMode append\n
-; data: settleDuration 500\n
 ; data: useViewTransition true\n
 ; data: fragments <div>hello</div>\n\n\n"
 
@@ -81,6 +81,5 @@
 ; retry: 1000\n
 ; data: selector foo\n
 ; data: mergeMode append\n
-; data: settleDuration 500\n
 ; data: useViewTransition true\n
 ; data: fragments <div>hello</div>\n\n\n"