+++ site/routes_tests.go
@@ -50,6 +50,7 @@ func setupTests(ctx context.Context, router chi.Router) (err error) {
 				{ID: "merge_fragment_whitespace"},
 				{ID: "on_load"},
 				{ID: "on_load_delay"},
+				{ID: "plugin_name_prefix"},
 				{ID: "radio_value"},
 				{ID: "ref"},
 				{ID: "remove_fragment"},
@@ -58,6 +59,7 @@ func setupTests(ctx context.Context, router chi.Router) (err error) {
 				{ID: "signals_change"},
 				{ID: "signals_change_path"},
 				{ID: "signals_change_path_once"},
+				{ID: "sse_events"},
 			},
 		},
 	}
@@ -118,6 +120,7 @@ func setupTests(ctx context.Context, router chi.Router) (err error) {
 			setupTestsMergeFragmentWhitespace(testsRouter),
 			setupTestsOnLoad(testsRouter),
 			setupTestsRemoveFragment(testsRouter),
+			setupTestsSseEvents(testsRouter),
 		); err != nil {
 			panic(fmt.Sprintf("error setting up tests routes: %s", err))
 		}