+++ site/smoketests/plugin_name_prefix_test.go
@@ -1,0 +1,9 @@
+package smoketests
+
+import (
+	"testing"
+)
+
+func TestUnitPluginNamePrefix(t *testing.T) {
+	setupPageTestOnLoad(t, "tests/plugin_name_prefix")
+}